WIP: Adding banner in hero-paid
